6354 | 15 files | M DumpSource2/schemas/client/CDOTA_Modifier_Broodmother_StickySnare.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Furbolg_Enrage_AttackSpeed_OnDeath.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Snapfire_SpitCreep_ArcingUnit.h,  M DumpSource2/schemas/client/CDOTA_Modifier_VoidSpirit_AetherRemnant_Unit.h,  M DumpSource2/schemas/client/CDOTA_Modifier_VoidSpirit_AetherRemnant_WatchThinker.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Broodmother_StickySnare.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Furbolg_Enrage_AttackSpeed_OnDeath.h,  M DumpSource2/schemas/server/CDOTA_Modifier_Snapfire_SpitCreep_ArcingUnit.h,  M DumpSource2/schemas/server/CDOTA_Modifier_VoidSpirit_AetherRemnant_Unit.h,  M DumpSource2/schemas/server/CDOTA_Modifier_VoidSpirit_AetherRemnant_WatchThinker.h,  M game/bin/built_from_cl.txt,  M game/dota/pak01_dir.txt,  M game/dota/readonly_tools_asset_info.txt,  M game/dota/steam.inf, ?? content/dota/materials/models/heroes/void_spirit/
https://steamdb.info/patchnotes/17729719/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Snapfire_SpitCreep_ArcingUnit.h
@@ -4,7 +4,7 @@ class CDOTA_Modifier_Snapfire_SpitCreep_ArcingUnit
 	float32 min_lob_travel_time;
 	float32 max_lob_travel_time;
 	float32 impact_radius;
-	int32 projectile_vision;
+	float32 projectile_vision;
 	float32 stun_duration;
 	float32 min_height_above_lowest;
 	float32 min_height_above_highest;
